# Conflicts:
#	main.c
#	minishell.h

+++ redirects.c
@@ -1,12 +1,41 @@
 #include "minishell.h"
 
-char	*get_file_name(char *str, int *i, t_all *all)
+char	*get_file_name(char *str, int *i, int type, t_all *all)
 {
 	int	len;
-	char *tmp;
+	char	*tmp;
+	char	*str_origin;
 	int j;
+	int x;
 
 	skip_spaces(str, i);
+
+	str_origin = ft_strdup(str);
+	x = *i; //чтобы в конце вернуть указатель после имени файла
+	while (str_origin[x] && !(check_set(str_origin[x], " \t|;<>")))
+	{
+		if (str[x] == '\'')
+		{
+			x++;
+			while (str[x] != '\'')
+				x++;
+			x++;
+		}
+		else if (str[x] == '\"')
+		{
+			x++;
+			while (str[x] != '\"')
+				x++;
+			x++;
+		}
+		else
+			x++;
+	}
+
+
+
+	if (type != 2)
+		str = replace_env_with_value(str, j, all);
 	len = get_arg_len(str, *i);
 	if (!len)
 		return NULL;
@@ -45,6 +74,7 @@ char	*get_file_name(char *str, int *i, t_all *all)
 		j++;
 	}
 	tmp[j] = '\0';
+	*i = x;
 	return(tmp);
 }
 
@@ -54,11 +84,12 @@ void	heredoc_stdin_read(t_all *all, char *stop)
 	char 	*heredoc_file;
 	int		ret;
 
-	heredoc_file = ft_strjoin(stop, "tmp_file");
+//	heredoc_file = ft_strjoin(stop, "tmp_file");
+	heredoc_file = ft_strdup("tmp_file_heredoc");
 	all->cmnd[all->pip_count].fd_in = open(heredoc_file, O_CREAT | O_RDWR | O_TRUNC, S_IRWXU);
 	if (all->cmnd[all->pip_count].fd_in < 0 || read(all->cmnd[all->pip_count].fd_in, 0, 0) < 0)
 	{
-		all->last_exit = errno;
+		g_status = errno;
 		printf("minishell: %s: %s\n", heredoc_file, strerror(errno));
 		unlink(heredoc_file);
 		ft_memdel(heredoc_file);
@@ -79,9 +110,10 @@ void	heredoc_stdin_read(t_all *all, char *stop)
 
 
 	close (all->cmnd[all->pip_count].fd_in);
-	all->cmnd[all->pip_count].fd_in = open("tmp_file", O_RDONLY);
+	all->cmnd[all->pip_count].fd_in = open(heredoc_file, O_RDONLY);
+	if (all->cmnd[all->pip_count].fd_in < 0 || read(all->cmnd[all->pip_count].fd_in, 0, 0) < 0)
 	{
-		all->last_exit = errno;
+		g_status = errno;
 		printf("minishell: %s: %s\n", heredoc_file, strerror(errno));
 		unlink(heredoc_file);
 		ft_memdel(heredoc_file);
@@ -100,11 +132,18 @@ void	ft_handle_redirect(char *str, int *i, t_all *all)
 	if (str[*i] == '>' && str[*i + 1] != '>')
 	{
 		(*i)++;
-		file_name = get_file_name(str, i, all);
+		file_name = get_file_name(str, i, 1, all);
+		if (!file_name)
+		{
+			g_status = 1;
+			printf("minishell: ambiguous redirect\n");
+			all->cmnd[all->pip_count].fd_out = -2;
+			return ;
+		}
 		all->cmnd[all->pip_count].fd_out = open(file_name, O_CREAT | O_RDWR | O_TRUNC, S_IRWXU);
 		if (all->cmnd[all->pip_count].fd_out < 0 || read(all->cmnd[all->pip_count].fd_out, NULL, 0) < 0)
 		{
-			all->last_exit = errno;
+			g_status = errno;
 			printf("minishell: %s: %s\n", file_name, strerror(errno));
 			ft_memdel(file_name);
 			return ;//todo maybe change function to int?
@@ -114,11 +153,18 @@ void	ft_handle_redirect(char *str, int *i, t_all *all)
 	else if (str[*i] == '>' && str[*i + 1] == '>')
 	{
 		*i = *i + 2;
-		file_name = get_file_name(str, i, all);
+		file_name = get_file_name(str, i, 1, all);
+		if (!file_name)
+		{
+			g_status = 1;
+			printf("minishell: ambiguous redirect\n");
+			all->cmnd[all->pip_count].fd_out = -2;
+			return ;
+		}
 		all->cmnd[all->pip_count].fd_out = open(file_name, O_CREAT | O_RDWR | O_APPEND, S_IRWXU);
 		if (all->cmnd[all->pip_count].fd_out < 0 || read(all->cmnd[all->pip_count].fd_out, NULL, 0) < 0)
 		{
-			all->last_exit = errno;
+			g_status = errno;
 			printf("minishell: %s: %s\n", file_name, strerror(errno));
 			ft_memdel(file_name);
 			return ;//todo maybe change function to int?
@@ -127,11 +173,18 @@ void	ft_handle_redirect(char *str, int *i, t_all *all)
 	else if (str[*i] == '<' && str[*i + 1] != '<')
 	{
 		(*i)++;
-		file_name = get_file_name(str, i, all);
+		file_name = get_file_name(str, i, 1, all);
+		if (!file_name)
+		{
+			g_status = 1;
+			printf("minishell: ambiguous redirect\n");
+			all->cmnd[all->pip_count].fd_in = -2;
+			return ;
+		}
 		all->cmnd[all->pip_count].fd_in = open(file_name, O_RDONLY);
 		if (all->cmnd[all->pip_count].fd_in < 0 || read(all->cmnd[all->pip_count].fd_in, NULL, 0) < 0)
 		{
-			all->last_exit = errno;
+			g_status = errno;
 			printf("minishell: %s: %s\n", file_name, strerror(errno));
 			ft_memdel(file_name);
 			return ;//todo maybe change function to int?
@@ -140,8 +193,8 @@ void	ft_handle_redirect(char *str, int *i, t_all *all)
 	else if (str[*i] == '<' && str[*i + 1] == '<')
 	{
 		*i = *i + 2;
-		stop_word = get_file_name(str, i, all);
-		heredoc_stdin_read(all, file_name);
+		stop_word = get_file_name(str, i, 2, all);
+		heredoc_stdin_read(all, stop_word);
 		ft_memdel(stop_word);
 	}
 }
